Refactored final blueprint, production
