+++ queertk/artist/views.py
@@ -1,9 +1,15 @@
 from flask import render_template
 from .artist import bp_artist
-from database import db
 
+# Import remote models
+from queertk.blueprints.common.models import Credit
+from queertk.production.models import Production
+
+# Import local models
 from .models import Artist
-from queertk.models import Credit, Production
+
+# Import database object
+from database import db
 
 @bp_artist.route('/<int:id>')
 @bp_artist.route('/<int:id>/<string:name>')
